Rolls back several changes that are preventing the CI from going green.
